Create useAirtableResources hook in utils
Move the related logic to that file / remove from resources.tsx per
comments on PR
